+++ app/Controllers/Home.php
@@ -2,10 +2,19 @@
 
 namespace App\Controllers;
 
+use CodeIgniter\CodeIgniter;
+
 class Home extends BaseController
 {
     public function index()
     {
-        return view('welcome_message');
+        $data = [
+            'CI_VERSION' => CodeIgniter::CI_VERSION,
+            'ENVIRONMENT' => ENVIRONMENT,
+            'YEAR' => date('Y'),
+        ];
+
+        // Render 'app\Views\welcome_message.mustache'
+        return $this->render('welcome_message.mustache', $data);
     }
 }