+++ CHANGELOG.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.0] - 2022-08-07
-
 ### Added
+
+-   Add Mustache Template Engine library dependencies.
+-   Add Mustache classes and render function to Base Controller.
+-   Add Mustache Filters to use in templating.
+
+### Changed
+
+-   Update Home Controller to use render function.
+-   Rename welcome_message.php to welcome_message.mustache.
+
+## [1.0.0] - 2022-08-07
 
 **Initial commit**